backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
boto3==1.9.248
botocore==1.12.248        # via boto3, s3transfer
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
coreapi==2.3.3            # via drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
django-cors-headers==3.1.1
django-extensions==2.2.3
django-nose==1.4.6
django==2.2.6
djangorestframework==3.10.3
docutils==0.15.2          # via botocore
drf-nested-routers==0.91
drf-yasg==1.17.0
idna==2.8                 # via requests
inflection==0.3.1         # via drf-yasg
itypes==1.1.0             # via coreapi
jinja2==2.10.3            # via coreschema
jmespath==0.9.4           # via boto3, botocore
markupsafe==1.1.1         # via jinja2
nose==1.3.7               # via django-nose
packaging==19.2           # via drf-yasg
psycopg2==2.8.3
pyparsing==2.4.2          # via packaging
python-dateutil==2.8.0    # via botocore
pytz==2019.3              # via django
requests==2.22.0          # via coreapi
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.5       # via drf-yasg
s3transfer==0.2.1         # via boto3
six==1.12.0               # via django-extensions, drf-yasg, packaging, python-dateutil
sqlparse==0.3.0           # via django
uritemplate==3.0.0        # via coreapi, drf-yasg
urllib3==1.25.6           # via botocore, requests
